requirements.txt
alembic==1.5.2
blinker==1.4
click==7.1.2
colorama==0.4.4
Deprecated==1.2.12
Flask==1.1.2
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
gunicorn==20.0.4
importlib-metadata==3.10.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
pendulum==2.1.2
Pillow==8.2.0
psycopg2==2.8.6
py-buzz==1.0.3
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytzdata==2020.1
qrcode==6.1
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.1
